+++ packages/@magic-sdk/types/CHANGELOG.md
@@ -1,3 +1,19 @@
+# v24.10.0 (Thu Oct 10 2024)
+
+#### 🚀 Enhancement
+
+- feat: add ping pong types for heartbeat [#815](https://github.com/magiclabs/magic-js/pull/815) ([@khamdam-magic](https://github.com/khamdam-magic))
+
+#### 🐛 Bug Fix
+
+- feat: add ping pong types ([@khamdam-magic](https://github.com/khamdam-magic))
+
+#### Authors: 1
+
+- [@khamdam-magic](https://github.com/khamdam-magic)
+
+---
+
 # v24.9.0 (Wed Sep 25 2024)
 
 #### 🚀 Enhancement